refactor: work on expr-eval integration
