+++ doc/examples/dynamic-content/conditional-expr.js
@@ -4,8 +4,6 @@ const title = 'Conditional content (expression)'
 
 const description = `The JSON schema if/then/else syntax can be cumbersome and dependencies too limited. This is why vjsf accepts the \`x-if\` annotation that can contain en evaluated expression. See the [Expressions documentation](configuration#expressions).
 
-To use rich expression with the expr-eval module, please se [rich-expression section](examples#rich-expression)
-
 **WARNING:** this syntax is entirely ignored by a JSON schema validator, therefore you can create a situation where the form is valid but the model is actually invalid (for example if you use \`x-if\` on a property that is also required).
 
 `
@@ -18,7 +16,14 @@ const schema = {
       title: 'I\'m a property rendered if the context contains a truthy value',
       'x-if': 'context.showActive'
     },
-    inactiveProp: { type: 'string', 'x-if': 'context.showInactive' }
+    inactiveProp: { type: 'string', 'x-if': 'context.showInactive' },
+    conditionProp: { type: 'string', title: `I'm a string whose content is used to evaluate an expression` },
+    conditionalProp: {
+      type: 'string',
+      title: `I'm here if previous prop is equal to "hello"`,
+      'x-if': 'parent.value.conditionProp  == \'hello\'',
+      'x-options': {evalMethod: 'evalExpr'}
+    }
   },
   oneOf: [{
     'x-if': 'context.showActive',
@@ -38,10 +43,12 @@ const schema = {
     properties: {
       type: { type: 'string', const: 'inactiveChoice' }
     }
-  }]
+  }],
 }
 
-const model = {}
+const model = {
+  conditionProp: 'hello'
+}
 
 const options = {
   context: {
