+++ lib/VJsfNoDeps.js
@@ -16,7 +16,7 @@ import MarkdownEditor from './mixins/MarkdownEditor'
 import Tooltip from './mixins/Tooltip'
 import Validatable from './mixins/Validatable'
 import Dependent from './mixins/Dependent'
-import Parser from './utils/rich-expression-parser'
+import exprEvalParser from './utils/expr-eval-parser'
 const expr = require('property-expr')
 
 const debugExpr = require('debug')('vjsf:expr')
@@ -343,6 +343,7 @@ export default {
 
       this._vjsf_getters = this._vjsf_getters || {}
 
+      // newFunction can only be defined on main options (not x-options to prevent injection)
       if (this.initialOptions.evalMethod === 'newFunction') {
         debugExpr(`evaluate expression "${exp}" with newFunction method`, expData)
         // use a powerful meta-programming approach with "new Function", not safe if the schema is user-submitted
@@ -351,9 +352,12 @@ export default {
         const result = this._vjsf_getters[exp](...Object.values(expData))
         debugExpr(`result`, result)
         return result
-      } else if (this.initialOptions.evalMethod === 'evalExpr') {
-        const parser = new Parser(require('expr-eval').Parser)
-        return parser.evaluate(exp, expData)
+      } else if (this.fullOptions.evalMethod === 'evalExpr') {
+        debugExpr(`evaluate expression "${exp}" with exprEval method`, expData)
+        // TODO: conserve compiled expression for reuse ?
+        const result = exprEvalParser.evaluate(exp, expData)
+        debugExpr(result)
+        return result
       } else {
         exp = this.prefixExpr(exp)
         debugExpr(`evaluate expression "${exp}" with propertyExpr method`, expData)
